requirements.txt
requests
fpdf2